Fix domain/data packages ambiguity
