+++ app/src/main/java/com/jaspersoft/android/jaspermobile/data/validator/SpringCredentialsValidator.java
@@ -24,12 +24,36 @@
 
 package com.jaspersoft.android.jaspermobile.data.validator;
 
-import com.jaspersoft.android.jaspermobile.data.server.JasperServer;
+import com.jaspersoft.android.jaspermobile.domain.BaseCredentials;
+import com.jaspersoft.android.jaspermobile.domain.validator.CredentialsValidator;
+import com.jaspersoft.android.sdk.service.auth.Credentials;
+import com.jaspersoft.android.sdk.service.auth.JrsAuthenticator;
+import com.jaspersoft.android.sdk.service.auth.SpringCredentials;
+
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
 /**
  * @author Tom Koptel
  * @since 2.3
  */
-public interface JasperServerValidator {
-    boolean validate(JasperServer instance);
+@Singleton
+public final class SpringCredentialsValidator implements CredentialsValidator {
+    private final JrsAuthenticator mAuthenticator;
+
+    @Inject
+    public SpringCredentialsValidator(JrsAuthenticator authenticator) {
+        mAuthenticator = authenticator;
+    }
+
+    @Override
+    public boolean validate(BaseCredentials credentials) {
+        Credentials springCredentials = SpringCredentials.builder()
+                .password(credentials.getPassword())
+                .username(credentials.getUsername())
+                .organization(credentials.getOrganization())
+                .build();
+        String token = mAuthenticator.authenticate(springCredentials);
+        return token != null;
+    }
 }
