+++ app/src/main/java/com/jaspersoft/android/jaspermobile/domain/server/JasperServerFactory.java
@@ -22,31 +22,12 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.jaspermobile.domain.entity.mapper;
-
-import android.support.annotation.NonNull;
-
-import com.jaspersoft.android.jaspermobile.data.server.JasperServer;
-import com.jaspersoft.android.sdk.service.data.server.ServerInfo;
-
-import javax.inject.Inject;
-import javax.inject.Singleton;
+package com.jaspersoft.android.jaspermobile.domain.server;
 
 /**
  * @author Tom Koptel
  * @since 2.3
  */
-@Singleton
-public class ServerInfoDataMapper {
-    @Inject
-    public ServerInfoDataMapper() {}
-
-    @NonNull
-    public JasperServer transform(String baseUrl, ServerInfo serverInfo) {
-        JasperServer.Builder builder = JasperServer.builder();
-        builder.setBaseUrl(baseUrl);
-        builder.setEdition(serverInfo.getEdition().name());
-        builder.setVersion(serverInfo.getVersion().getVersionCode());
-        return builder.create();
-    }
+public interface JasperServerFactory {
+    JasperServer create(String baseUrl);
 }
