+++ app/src/main/java/com/jaspersoft/android/jaspermobile/domain/validator/ProfileValidator.java
@@ -24,36 +24,12 @@
 
 package com.jaspersoft.android.jaspermobile.domain.validator;
 
-import com.jaspersoft.android.jaspermobile.data.BaseCredentials;
-import com.jaspersoft.android.jaspermobile.data.validator.CredentialsValidator;
-import com.jaspersoft.android.sdk.service.auth.Credentials;
-import com.jaspersoft.android.sdk.service.auth.JrsAuthenticator;
-import com.jaspersoft.android.sdk.service.auth.SpringCredentials;
-
-import javax.inject.Inject;
-import javax.inject.Singleton;
+import com.jaspersoft.android.jaspermobile.domain.Profile;
 
 /**
  * @author Tom Koptel
  * @since 2.3
  */
-@Singleton
-public final class SpringCredentialsValidator implements CredentialsValidator {
-    private final JrsAuthenticator mAuthenticator;
-
-    @Inject
-    public SpringCredentialsValidator(JrsAuthenticator authenticator) {
-        mAuthenticator = authenticator;
-    }
-
-    @Override
-    public boolean validate(BaseCredentials credentials) {
-        Credentials springCredentials = SpringCredentials.builder()
-                .password(credentials.getPassword())
-                .username(credentials.getUsername())
-                .organization(credentials.getOrganization())
-                .build();
-        String token = mAuthenticator.authenticate(springCredentials);
-        return token != null;
-    }
+public interface ProfileValidator {
+    boolean validate(Profile profile);
 }
