+++ app/src/test/java/com/jaspersoft/android/jaspermobile/domain/ProfileTest.java
@@ -22,7 +22,7 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.jaspermobile.data;
+package com.jaspersoft.android.jaspermobile.domain;
 
 import org.junit.Test;
