+++ app/src/test/java/com/jaspersoft/android/jaspermobile/domain/server/JasperServerTest.java
@@ -22,44 +22,19 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.jaspermobile.domain.validator;
+package com.jaspersoft.android.jaspermobile.domain.server;
 
-import com.jaspersoft.android.jaspermobile.data.server.JasperServer;
-
-import org.junit.Before;
 import org.junit.Test;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
+import nl.jqno.equalsverifier.EqualsVerifier;
 
 /**
  * @author Tom Koptel
  * @since 2.3
  */
-public class JasperServerValidatorImplTest {
-
-    JasperServerValidatorImpl validator;
-
-    @Before
-    public void setUp() throws Exception {
-        validator = new JasperServerValidatorImpl();
-    }
-
+public class JasperServerTest {
     @Test
-    public void serverThatIsEquals5_5IsValid() throws Exception {
-        boolean isValid = validator.validate(JasperServer.builder().setVersion(5.5d).create());
-        assertThat(isValid, is(true));
-    }
-
-    @Test
-    public void serverThatIsHigherThan5_5IsValid() throws Exception {
-        boolean isValid = validator.validate(JasperServer.builder().setVersion(6.0d).create());
-        assertThat(isValid, is(true));
-    }
-
-    @Test
-    public void serverThatIsEquals5_0IsNotValid() throws Exception {
-        boolean isValid = validator.validate(JasperServer.builder().setVersion(5.0d).create());
-        assertThat(isValid, is(false));
+    public void equalsContract() {
+        EqualsVerifier.forClass(JasperServer.class).verify();
     }
 }
